Removed backwards compatibility for legacy mainActivity + update map pins on story delete
