+++ app/src/main/AndroidManifest.xml
@@ -35,7 +35,7 @@
         android:usesCleartextTraffic="true">
 
         <activity
-            android:name=".core.composables.StorySpotsMainActivity"
+            android:name=".MainActivity"
             android:exported="true"
             android:theme="@style/Theme.StorySpots">
         </activity>
