--- app/src/main/java/com/storyspots/core/composables/MainScreen.kt
@@ -1,4 +1,0 @@
-package com.storyspots.core.composables
-
-class MainScreen {
-}
