--- app/src/main/java/com/storyspots/core/composables/StorySpotsMainActivity.kt
@@ -1,81 +1,0 @@
-package com.storyspots.core.composables
-
-import android.os.Bundle
-import android.widget.Toast
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
-import com.mapbox.android.core.permissions.PermissionsListener
-import com.mapbox.android.core.permissions.PermissionsManager
-import com.storyspots.core.AppComponents
-import com.storyspots.ui.StorySpotsApp
-import kotlinx.coroutines.launch
-
-class StorySpotsMainActivity : ComponentActivity(), PermissionsListener {
-
-    private var permissionsManager: PermissionsManager? = null
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        // Set content immediately - don't wait for anything
-        setContent {
-            StorySpotsApp()
-        }
-
-        // Handle permissions asynchronously
-        lifecycleScope.launch {
-            lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
-                checkAndRequestPermissions()
-            }
-        }
-    }
-
-    private suspend fun checkAndRequestPermissions() {
-        if (!PermissionsManager.areLocationPermissionsGranted(this)) {
-            permissionsManager = PermissionsManager(this)
-            permissionsManager?.requestLocationPermissions(this)
-        } else {
-            AppComponents.permissionManager.setLocationPermissionGranted(true)
-        }
-    }
-
-    @Deprecated("This method has been deprecated in favor of using the Activity Result API\n      which brings increased type safety via an {@link ActivityResultContract} and the prebuilt\n      contracts for common intents available in\n      {@link androidx.activity.result.contract.ActivityResultContracts}, provides hooks for\n      testing, and allow receiving results in separate, testable classes independent from your\n      activity. Use\n      {@link #registerForActivityResult(ActivityResultContract, ActivityResultCallback)} passing\n      in a {@link RequestMultiplePermissions} object for the {@link ActivityResultContract} and\n      handling the result in the {@link ActivityResultCallback#onActivityResult(Object) callback}.")
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<String>,
-        grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        permissionsManager?.onRequestPermissionsResult(requestCode, permissions, grantResults)
-    }
-
-    override fun onExplanationNeeded(permissionsToExplain: List<String>) {
-        Toast.makeText(
-            this,
-            "Location permission is needed to show your position on the map",
-            Toast.LENGTH_LONG
-        ).show()
-    }
-
-    override fun onPermissionResult(granted: Boolean) {
-        AppComponents.permissionManager.setLocationPermissionGranted(granted)
-        if (!granted) {
-            Toast.makeText(
-                this,
-                "Location features will be limited",
-                Toast.LENGTH_SHORT
-            ).show()
-        }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        if (isFinishing) {
-            AppComponents.mapManager.cleanup()
-            AppComponents.cleanup()
-        }
-    }
-}
