+++ app/src/main/java/com/storyspots/register/RegisterActivity.kt
@@ -7,11 +7,9 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.storyspots.MainActivity
-import com.storyspots.core.composables.StorySpotsMainActivity
-import com.storyspots.login.LoginActivity
 
 class RegisterActivity : ComponentActivity() {
-    private val intentClass = StorySpotsMainActivity::class.java
+    private val intentClass = MainActivity::class.java
 //    private val intentClass = MainActivity::class.java
 
     override fun onCreate(savedInstanceState: Bundle?) {
